Don't fail user creation with a mysterious "constraint violation" error
If an admin is trying to create a user whose username is already
in use in some other school that they cannot access, the system
would crash with a cryptic "constraint violation" error. Detect
and hack around this error and provide a proper error message.

This also applies to existing user editing, if the new username
is already in use.