Remove sassc-rails gem and let ESBuild handle CSS bundling
I had to change the name of the bundles directory and the CSS
bundle, because ESBuild refused to cooperate. I'll take another
look at it later.

All tests pass and "make deb" works fine and the resulting
.deb packages install just fine and they produce a working
puavo-standalone environment (I actually tested this).
But I'm still prepared to revert this commit, because so
many things can go wrong.