Initial token auditing implementation
Repetitive and ugly code, but it works.

Keeping a history of issued tokens and their uses is a bit controversial,
but it's not enabled by default. I know that sooner or later, we'll have
to go through these logs when dealing with abuse cases.

Logging IP addresses is even more controlversial, as it makes us susceptible
to GDPR issues. Which is why it isn't enabled by default. It can be enabled
by default once we have a system for deleting/scrubbing old records (basically
deleting rows or clearing out the IP address column if the rows are older
than N days).