Permit logins to service using another service's session
This was a nice idea, except... it does not work in production.
Some of our organisations really do have multiple services which
all want to share the same session (emphasizing the *single* part
of "single-sign on").

Perhaps this could be a per-organisation (and per-service?) setting
instead? Those who want the extra security could get it.