Add a menu-driven OAuth2 token manager script
I've been working on this since January 2025. It works, but boy it is a mess.
A cleaner web-based frontend is required, but currently there's no time for
it yet, as the system is still in motion and large architectural changes
are possible. This CLI script is easier to update than a complex web page.