Rename the Opinsys employees feature to super owners
Super owners are owners who have been granted even more permissions. The
list of super owners is kept on the server, because if it was stored in
LDAP, who would have a write access to it? Not saying a file on the disk
is the best way, but currently it's the best way we can do it.