Reduce startup fragility
For unknown reasons, if the file cannot be loaded (it's missing,
or there's no OAuth2 configuration, for example) the exception
can somehow "leak" and cause puma to enter a crash loop at startup.

I can reproduce this in a production environment, but not in a
development environment. Very strange. This check seems to fix
the most acute problem, though.