Fix User model constructor.
ActiveLdap can use a model so it gives a DN as an argument to it.
I suppose our way of overriding the constructor is wrong, perhaps?
But I do not know what else to do, so check that the argument is
a hash, when setting the default value for puavoTeacherPermissions,
and in case it is not, just pass the argument to the parent
constructor.