Toolbar and dropdown builder helper class
The purpose of this class is to make it easier to specify toolbars and
dropdown menus on various index and show pages.

Currently the toolboxes are just plain UL/LI lists, mixed with styles
and icons and other messy stuff. This class is an alternative to all
that.

Many of the new CSS class names have "2" in them to avoid the styles
from the existing toolbox classes from leaking into them. Once the
old toolboxes have been removed, the CSS classes can be renamed.

I first sketched this back in November 2022. That shows how far behind
I'm on my TODO list, and it's not even the oldest thing on the list.