Fix external file saving
Do you see the bug? It's one missing "]" character. The bug has been in
puavo-web since commit 4eb2699. But it has worked just fine for years,
until some recent Rails upgrades finally broke it.

The bug caused a "[" character to be added to the file name when the
external file was saved. So its name no longer matched to anything
listed in puavo_external_files.yml and you couldn't see any changes
in the UI.

And while fixing this I realized we have no puavo-web tests for external
files. There are some tests in puavo-rest, but they won't see this.